+++ src/main/java/com/vuthy/mobilebankingapi/service/SegmentService.java
@@ -2,6 +2,8 @@ package com.vuthy.mobilebankingapi.service;
 
 import com.vuthy.mobilebankingapi.domain.Segment;
 
+import java.math.BigDecimal;
+
 public interface SegmentService {
-    Segment createSegment(String name, Integer overLimit);
+    Segment createSegment(String name, BigDecimal overLimit);
 }
